убрал лишнее и сгруппировал
убрал лишнее и сгруппировал
