--- routers/commands/__init__.py
@@ -1,11 +1,0 @@
-__all__ = ("router",)
-
-from aiogram import Router
-
-from .base_commands import router as base_commands_router
-
-router = Router(name=__name__)
-
-router.include_routers(
-    base_commands_router
-)
